requirements.txt
discord
asyncio
urllib3
requests
aiohttp
bs4
nltk
unidecode
aiodns
cchardet
colorama
lomond
websocket-client
websockets
names
pymongo
sendgrid
tweepy
paypalrestsdk
